Grelease: v2.3.0
First official public release of bgscan.

This commit includes all accumulated development work up to version 2.3.0.

Originally developed privately, the project is now being published to help the community scan OpenConnect endpoints and support anti-censorship efforts.

Major components included:
- Modular multi-stage scanning engine
- Bubble Tea interactive TUI
- Advanced protocol chaining
- Xray / DNS / Slipstream integration
- Asset system
- Structured settings system
- Cross-platform build system

Future commits will follow a structured incremental workflow."

+++ internal/core/xray/inbound.go
@@ -10,7 +10,7 @@ package xray
 // Sniffing is enabled with HTTP and TLS destination overrides so that
 // Xray can correctly detect protocols even when the target service
 // does not explicitly specify them.
-func getInbound(port int) Inbound {
+func getInbound(port uint16) Inbound {
 	return Inbound{
 		Port:     port,
 		Listen:   "127.0.0.1",
